+++ src/main/java/org/mangorage/render/core/registry/IHolder.java
@@ -2,7 +2,7 @@ package org.mangorage.render.core.registry;
 
 public interface IHolder<T> {
     T getValue();
-    byte getId();
+    int getId();
 
     default boolean is(IHolder<?> holder) {
         return holder.getValue() == getValue();
